Finish moving Netoworking strategies to new files. Clean up folder structure.
