--- BedrockService/Service/Configs/Default.conf
@@ -1,28 +1,0 @@
-# Server
-server-name=Default
-gamemode=creative
-difficulty=easy
-allow-cheats=false
-max-players=10
-online-mode=true
-white-list=false
-server-port=19132
-server-portv6=19133
-view-distance=32
-tick-distance=4
-player-idle-timeout=30
-max-threads=8
-level-name=Bedrock level
-level-seed=
-default-player-permission-level=member
-texturepack-required=false
-content-log-file-enabled=false
-compression-threshold=1
-server-authoritative-movement=server-auth
-player-movement-score-threshold=20
-player-movement-distance-threshold=0.3
-player-movement-duration-threshold-in-ms=500
-correct-player-movement=false
-
-# StartCmds
-AddStartCmd=help 1
