--- BedrockService/Service/Configs/Globals.conf
@@ -1,16 +1,0 @@
-#Globals
-ServersPath=C:\MCBedrockService
-AcceptedMojangLic=false
-ClientPort=19134
-#Backups
-BackupEnabled=false
-BackupPath=Default
-BackupCron=0 1 * * *
-MaxBackupCount=25
-EntireBackups=false
-#Updates
-CheckUpdates=true
-UpdateCron=38 19 * * *
-#Logging
-LogServersToFile=false
-LogServiceToFile=false
