+++ BedrockService/Service/Core/BedrockService.cs
@@ -1,6 +1,7 @@
-﻿using BedrockService.Service.Core.Interfaces;
+﻿using BedrockService.Service.Management.Interfaces;
+using BedrockService.Service.Networking.Interfaces;
 using BedrockService.Service.Server;
-using BedrockService.Shared.Utilities;
+using BedrockService.Service.Server.Interfaces;
 using NCrontab;
 using System.Timers;
