+++ BedrockService/Service/Networking/Interfaces/IFlaggedMessageParser.cs
@@ -1,4 +1,4 @@
-﻿namespace BedrockService.Service.Networking.MessageInterfaces {
+﻿namespace BedrockService.Service.Networking.Interfaces {
     public interface IFlaggedMessageParser {
         (byte[] data, byte srvIndex, NetworkMessageTypes type) ParseMessage(byte[] data, byte serverIndex, NetworkMessageFlags flag);
     }
