+++ BedrockService/Service/Networking/Interfaces/IMessageParser.cs
@@ -1,4 +1,4 @@
-﻿namespace BedrockService.Service.Networking.MessageInterfaces {
+﻿namespace BedrockService.Service.Networking.Interfaces {
     public interface IMessageParser {
         (byte[] data, byte srvIndex, NetworkMessageTypes type) ParseMessage(byte[] data, byte serverIndex);
     }
