+++ BedrockService/Service/Networking/NetworkStrategies/LevelEditFile.cs
@@ -1,0 +1,27 @@
+﻿using BedrockService.Service.Networking.Interfaces;
+
+namespace BedrockService.Service.Networking.NetworkStrategies {
+    public class LevelEditFile : IMessageParser {
+        private readonly IServiceConfiguration _serviceConfiguration;
+        private readonly IBedrockService _bedrockService;
+        private readonly IBedrockLogger _logger;
+
+        public LevelEditFile(IServiceConfiguration serviceConfiguration, IBedrockService bedrockService, IBedrockLogger logger) {
+            _logger = logger;
+            _bedrockService = bedrockService;
+            _serviceConfiguration = serviceConfiguration;
+        }
+
+        public (byte[] data, byte srvIndex, NetworkMessageTypes type) ParseMessage(byte[] data, byte serverIndex) {
+            byte[] stripHeaderFromBuffer = new byte[data.Length - 5];
+            Buffer.BlockCopy(data, 5, stripHeaderFromBuffer, 0, stripHeaderFromBuffer.Length);
+            IServerConfiguration server = _serviceConfiguration.GetServerInfoByIndex(serverIndex);
+            string pathToLevelDat = $@"{_serviceConfiguration.GetProp("ServersPath")}\{server.GetProp("server-name")}\worlds\{server.GetProp("level-name")}\level.dat";
+            _bedrockService.GetBedrockServerByIndex(serverIndex).AwaitableServerStop(false).Wait();
+            File.WriteAllBytes(pathToLevelDat, stripHeaderFromBuffer);
+            _logger.AppendLine($"level.dat writen to server {server.GetServerName()}");
+            _bedrockService.GetBedrockServerByIndex(serverIndex).AwaitableServerStart().Wait();
+            return (Array.Empty<byte>(), 0, 0);
+        }
+    }
+}
