+++ BedrockService/Service/Server/BedrockServer.cs
@@ -1,6 +1,6 @@
-﻿using BedrockService.Service.Server.Management;
+﻿using BedrockService.Service.Management.Interfaces;
+using BedrockService.Service.Server.Interfaces;
 using BedrockService.Shared.PackParser;
-using BedrockService.Shared.Utilities;
 using System.IO.Compression;
 
 namespace BedrockService.Service.Server {
